+++ src/Attendee.ts
@@ -1,4 +1,4 @@
-type Email = string;
-type Name = string;
+export type Email = string;
+export type Name = string;
 
 export type Attendee = Email | [Email, Name?];